app/backend/requirements.txt
Flask==3.1.1
colorama==0.4.6
fasttext==0.9.3
faiss-cpu
joblib==1.4.2
kneed==0.8.5
numpy<2.0.0,>=1.18.5
ollama==0.5.3
openai==1.106.1
pandas==2.3.2
protobuf>=4.25.8
pyarrow==16.1.0
openpyxl
python-dotenv==1.1.1
PyYAML==6.0.2
Requests==2.32.5
scikit_learn==1.7.1
sentence_transformers==5.1.0
tqdm==4.67.1
transformers==4.41.2
sentencepiece==0.2.1
sacremoses
datasets
google-genai>=1.0.0
langdetect
py7zr
pyLDAvis
